requirements.txt
numpy>=1.21.0
pandas>=1.3.0
gymnasium>=0.28.1
torch>=1.9.0
yfinance>=0.2.0
robin_stocks>=3.0.0
python-dotenv>=0.19.0
pyotp>=2.6.0
